Chatbot/requirements.txt
flask
openai
python-dotenv
langchain_community
flask-cors